requirements.txt
astronomer-cosmos>=1.0.2
apache-airflow-providers-snowflake